DAC Output Stablized

+++ ADIO_FINAL/Core/Inc/motor_driver.h
@@ -30,7 +30,7 @@ typedef enum {
 } Motor_ID_t;
 
 void Motor_Init(void);
-void Motor_Set(Motor_ID_t motor, float command);
+void Motor_Set(Motor_ID_t motor, uint16_t  command);
 void Motor_Stop(Motor_ID_t motor);
 void Motor_StopAll(void);
